pyproject.toml
[tool.poetry]
name = "hackaton-belcorp"
version = "0.1.0"
description = ""
authors = ["AmaruEscalante <amaruescalante@gmail.com>"]
readme = "README.md"
packages = [{include = "hackaton_belcorp"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
httpx = "^0.25.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
